refactor: extract bffMatchesResumeFilters to shared utility for testability (#801)
Extract bffMatchesResumeFilters and parseAgeFromContentField from
routes/resumes.ts into services/bff-filter-utils.ts so the filter
logic can be imported directly by tests instead of being replicated
inline. Rewrite bff-filter-alignment.test.ts to import the real
function, eliminating test-code drift risk. Add test coverage for
minRoleYears, age (minAge/maxAge), and location filters.
